Catch up with -current.

+++ sys/dev/vme/xd.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: xd.c,v 1.37 2001/07/22 13:34:10 wiz Exp $	*/
+/*	$NetBSD: xd.c,v 1.37.2.1 2001/10/01 12:46:39 fvdl Exp $	*/
 
 /*
  *
@@ -1627,7 +1627,7 @@ xdc_startbuf(xdcsc, xdsc, bp)
 	/* init iorq and load iopb from it */
 	xdc_rqinit(iorq, xdcsc, xdsc, XD_SUB_NORM | XD_MODE_VERBO, block,
 		   bp->b_bcount / XDFM_BPS,
-		   (caddr_t)iorq->dmamap->dm_segs[0].ds_addr,
+		   (caddr_t)(u_long)iorq->dmamap->dm_segs[0].ds_addr,
 		   bp);
 
 	xdc_rqtopb(iorq, iopb, (bp->b_flags & B_READ) ? XDCMD_RD : XDCMD_WR, 0);
